rename(file): change file path and name
