+++ apps/client/src/app/my-page/layout.tsx
@@ -1,5 +1,6 @@
-import { CommonLayout } from "@/components/layout/CommonLayout";
-import React from "react";
+import { CommonLayout } from '@/components/layout/CommonLayout';
+import { cookies } from 'next/headers';
+import React from 'react';
 
 interface Props {
   children: React.ReactNode;
