+++ apps/client/src/app/my-page/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ProfileContainer from '@/containers/auth/mypage/Profile';
+import ProfileContainer from '@/containers/my-page/Profile';
 
 const MyPage: React.FC = () => {
   return <ProfileContainer />;
