+++ apps/client/src/components/serviceinfo/PrivacyPolicy.tsx
