+++ apps/client/src/components/serviceinfo/TermsOfService.tsx
